+++ src/type-guards.ts
@@ -1,4 +1,4 @@
-import {IUseMixins, Mixin, WithMixin} from "./types";
+import {IMixinBase, IUseMixins, Mixin, MixinFull, MixinThis, WithMixin} from "./types";
 
 
 export function haveMixin<M extends Mixin<any, any>>(v: any, mixin: M): v is WithMixin<M> {
@@ -20,4 +20,8 @@ export function haveMixins<Mixins extends Array<Mixin<any, any>>>(v: any, ...mix
     }
 
     return true;
+}
+
+export function isMixin<M extends IMixinBase<any> = IMixinBase<any>>(value: any): value is M {
+    return typeof value === 'object' && 'mixinName' in value && 'target' in value;
 }